intermediate version
Saving this now to preserve some code that I'm about to toss out.  This
adds code for the rotation of the corner-colored remap into the standard
orientation, defined by the centers (instead of by the GRW corner).  All
of the maps that can map from the input corner coloring to the standard
one are utilized, and only minimum one after standard orientation is
kept.  The "minimum" form is defined by alphabetical ordering of a
serialization of the color characters of the cube.

Also:
- Move corner coloring code to its own file.